review1 change
The fully_conn function: 
there is a slight modification that needs to take place to improve the results. Remember back in Project 1, when a sigmoid function was used as an activation to make the network non-linear? The same thing needs to happen here.

The print_stats function:
But you are measuring the training accuracy instead of the validation accuracy. Use valid_features and valid_labels instead!